refactor: split currency conversion logic from Lua integration
Split the currency conversion implementation into two files:
- src/currency.rs: Core CurrencyConverter with all conversion logic, caching, and backend
- src/commands/execute/extensions/currency.rs: Lua integration that wraps the converter

Create a single CurrencyConverter instance at the top level in main.rs and pass it to:
- Lua extension for currency.convert(), currency.rate(), and currency.clear_cache()
- Discord slash command for /currency

This architecture:
- Eliminates code duplication
- Ensures both integrations use the same cache
- Makes the converter easily testable
- Follows dependency injection pattern
