+++ src/commands/execute/extensions/mod.rs
@@ -1,6 +1,7 @@
 use std::sync::Arc;
 
 use crate::ai::Ai;
+use crate::currency::CurrencyConverter;
 
 pub mod currency;
 mod globals;
@@ -10,12 +11,13 @@ mod perchance;
 pub fn register(
     lua: &mlua::Lua,
     ai: Arc<Ai>,
+    currency_converter: Arc<CurrencyConverter>,
     output_tx: flume::Sender<String>,
     print_tx: flume::Sender<String>,
 ) -> mlua::Result<()> {
     globals::register(lua, output_tx, print_tx)?;
     llm::register(lua, ai)?;
     perchance::register(lua)?;
-    currency::register(lua)?;
+    currency::register(lua, currency_converter)?;
     Ok(())
 }
